move importdata in to helpers
