+++ packages/fake-data/src/populateDb/index.ts
@@ -1,4 +1,4 @@
 export * from './generateEachDataType';
-export * from './generateImportData';
+export * from './helpers/importData';
 
 export * from './helpers';
